agent/requirements.txt
instagrapi>=2.0.0
yt-dlp>=2024.0.0
openai-whisper>=20231117
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
